Created sameElem.cpp
1.Sorting Arrays Approach
2.Using Unordered_set
3.Using Ordered set
Each approach has its own advantages and disadvantages depending on factors such as the size of the arrays and the distribution of elements. In general, using unordered set provides a more efficient solution in terms of time complexity compared to the other two approaches. However, the choice of approach may also depend on other factors such as space complexity and the need for a sorted result.