Create shrtCmnSuprS.cpp
Find the Longest Common Subsequence (LCS) of two given strings. For example, lcs of “geek” and “eke” is “ek”. 
Insert non-lcs characters (in their original order in strings) to the lcs found above, and return the result. So “ek” becomes “geeke” which is shortest common supersequence.
Let us consider another example, s1 = “AGGTAB” and s2 = “GXTXAYB”. LCS of s1 and s2 is “GTAB”. Once we find LCS, we insert characters of both strings in order and we get “AGXGTXAYB”